requirements.txt
annotated-types>=0.6.0
anyio>=4.3.0
dnspython>=2.6.1
email-validator>=2.1.0
fastapi>=0.109.0
fastui>=0.5.0
h11>=0.14.0
h2>=4.1.0
hpack>=4.0.0
hypercorn>=0.16.0
hyperframe>=6.0.1
idna>=3.6
openbb>=4.4.2
openbb-akshare==0.4.18
priority>=2.0.0
pydantic>=2.5.0
sniffio>=1.3.0
starlette>=0.36.0
typing-extensions>=4.8.0
wsproto>=1.2.0
python-dotenv>=1.0.0
openbb-yfinance==1.5.0
baostock==00.8.90
alpha_vantage>=2.3.1